PR2_updated: 1. Removed unwanted header files 2. Renamed heartbeat.xx files to heartbeat_task.xx 3. Renamed heartbeat_logic.xx files to heartbeat.xx 4. Did other formatting and added additional comments 5. Added paths to header file inclusions, relative to user/ 6. Renamed some functions
