+++ cli_ftd/cli.c
@@ -73,6 +73,7 @@
 #include "task_config.h"
 #include "ti_drivers_config.h"
 
+
 #if TIOP_OAD
 /* OAD required Header files */
 #include "oad_image_header.h"
