+++ cli_ftd/user/sensor/stm32/stm32_i2c.h
@@ -4,10 +4,10 @@
 #include <inttypes.h>
 #include <stdbool.h>
 
-#include <string.h>
-
+/* Initialization function for I2C slave */
 void stm32__init(void);
 
+/* Getter functions for sensor data */
 uint32_t stm32__spec_co_gas_concentration(void);
 int16_t stm32__spec_co_temperature(void);
 uint16_t stm32__spec_co_humidity(void);
@@ -26,14 +26,12 @@ void stm32__magnetometer(int32_t arr[]);
 void stm32__heartbeat_data(uint16_t arr[]);
 void stm32__gps_data(float data_arr[]);
 
-// Helper functions
+/* Helper functions */
 bool stm32__write_data(uint8_t write_address, uint8_t data);
 bool stm32__read_data(uint8_t write_address, uint8_t readBuf[],
                       uint32_t readCount);
 
-/**
- * Creation function for the stm32_i2c application task.
- */
+/* Creation function for the stm32_i2c application task */
 void stm_i2c_comm_taskCreate(void);
 
 #endif
