+++ cli_ftd/user/sensor/stm32/stm32_i2c_task.h
@@ -1,6 +1,7 @@
 #ifndef USER_STM32_I2C_SLAVE_H_
 #define USER_STM32_I2C_SLAVE_H_
 
+// Task creation function for I2C slave
 void stm_i2c_comm_taskCreate(void);
 
 #endif
