+++ cli_ftd/user/udp_communication/udp_comm.h
@@ -1,5 +1,7 @@
 #ifndef USER_UDP_COMM_H_
 #define USER_UDP_COMM_H_
 
-extern void udp_comm_taskCreate();
+/* Creation function for the UDP application task */
+void udp__comm_taskCreate();
+
 #endif
